% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-catalog.R
\docType{data}
\name{catalog}
\alias{catalog}
\title{e-stat statistics catalog data}
\format{A [tbl_df] with 1604 observations, unnested 1642 observations.}
\usage{
catalog
}
\description{
e-stat statistics catalog data
}
\section{Dimensions}{
 

\itemize{
  \item `title`: statistic survey title.
  \item `start_date`, `end_date` data recored periods. 
  \item `information`: data summary. nested data.frame
}
}

\examples{
catalog

tidyr::unnest(catalog)
}
