% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_jyr.R
\name{convert_jyr}
\alias{convert_jyr}
\title{Convert Japanese imperial year to Anno Domini}
\usage{
convert_jyr(jyr)
}
\arguments{
\item{jyr}{Japanese imperial year. kanji or roman character}
}
\description{
Convert Japanese imperial year to Anno Domini
}
\references{
Nippon wareki2AD
}
