% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_download.R
\name{download_estat_excel}
\alias{download_estat_excel}
\title{Download specific estat files using selenium driver}
\usage{
download_estat_excel(driver, url)
}
\arguments{
\item{driver}{RSelenium remoteDriver object}

\item{url}{Stats data information url (character)}
}
\description{
Download specific estat files using selenium driver
}
\examples{
\dontrun{
require("RSelenium")

crmDv <- wdman::chrome(port = 4445L)
crmDv$output()
remDr <- remoteDriver(port = 4445L,
browserName = "phantomjs")

download_estat_excel(remDr, url = "https://www.e-stat.go.jp/stat-search/files?\\n
page=1&layout=datalist&toukei=00500226&tstat=000001015623\\n
&cycle=7&tclass1=000001032194&tclass2=000001100735&stat_infid=000031571891")

eCaps <- list(chromeOptions = list(prefs = list(
"profile.default_content_settings.popups" = 0L,
"download.prompt_for_download" = FALSE,
"download.default_directory" = getwd())))

remDr <- remoteDriver(
port = 4445L,
browserName = "phantomjs",
extraCapabilities = eCaps)

download_estat_excel(remDr, url = "https://www.e-stat.go.jp/stat-search/files?\\n
page=1&layout=datalist&toukei=00500226&tstat=000001015623&cycle=7\\n
&tclass1=000001032194&tclass2=000001100735&stat_infid=000031571891")
}
}
